SWE_Gr14_Backend/requirements.txt
psycopg2-binary==2.9.10
python-dotenv
cryptography
bcrypt
pycryptodome